Tighten multimodal embedding public surface and docs
- `embedding_modality()` moves to a private `_modality` module; de-listing it
  from `__all__` didn't hide it, as mkdocstrings filters by name here
- `prepare_embed()` documents that an item may be an `EmbeddingContent`, which
  `supported_modalities` can't refuse
- docs record the trust model for embedding file URLs, which the server fetches
  itself, and the provider facts the new cassettes settled: the 6-image ceiling
  is per input, Vertex takes one input per request, and usage is 0 on the
  Gemini API because the SDK drops the token count the API returns
