+++ pydantic_ai_slim/pydantic_ai/embeddings/input.py
@@ -2,8 +2,6 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 from typing import Literal, TypeAlias
 
-from typing_extensions import assert_never
-
 from pydantic_ai.exceptions import UserError
 from pydantic_ai.messages import AudioUrl, BinaryContent, DocumentUrl, ImageUrl, TextContent, VideoUrl
 
@@ -78,28 +76,3 @@ EmbeddingInput: TypeAlias = EmbeddingContentPart | EmbeddingContent
 def embedding_parts(item: EmbeddingInput) -> Sequence[EmbeddingContentPart]:
     """The content parts that make up an input, so a single part and an `EmbeddingContent` can be handled alike."""
     return item.content if isinstance(item, EmbeddingContent) else [item]
-
-
-def embedding_modality(part: EmbeddingContentPart) -> EmbeddingModality:
-    """The modality of a content part, to check it against a model's supported modalities."""
-    if isinstance(part, str | TextContent):
-        return 'text'
-    elif isinstance(part, ImageUrl):
-        return 'image'
-    elif isinstance(part, AudioUrl):
-        return 'audio'
-    elif isinstance(part, VideoUrl):
-        return 'video'
-    elif isinstance(part, DocumentUrl):
-        return 'document'
-    elif isinstance(part, BinaryContent):
-        if part.is_image:
-            return 'image'
-        elif part.is_audio:
-            return 'audio'
-        elif part.is_video:
-            return 'video'
-        else:
-            return 'document'
-    else:
-        assert_never(part)
